Simplify consumer error handling.
I started looking into this because I noticed that when our postgres
database goes away (during mass reboots), datanommer goes into a corrupt
state *but* the nagios alarms about exception counts don't go off.  It
just errors out on every message coming in and we never know about it
until sysadmin-datanommer members check their email.

The code in moksha that calls this consume method has that exception
counting code, but datanommer here is trying to do the error logging
itself here.  By removing all of that and re-raising the exception,
moksha should correctly both send the email *and* notify nagios and
collectd of the exception.

Note that, there's all extra stuff here that's getting removed about
trying to re-save the message without a msg_id.  That was some extra
paranoid stuff from myself and ianweller, since we thought there might
be uuid conflicts.  In the 3 years datanommer's been running we haven't
hit that codepath once.  The UUIDs won't conflict -- they're prefixed
with the year.  The probability is tiny, so, simplify it.