use timedelta instead of datetime.replace()
datetime.replace() will raise a ValueError if the value you specify is
outside the valid range for that field. In this case, this test would
fail if run on the first or second day of the month, because it would
try to set day to 0 or -1. Subtracting a timedelta from a datetime
handles rollover correctly.