models.add: use INSERT ... SELECT to avoid errors from duplicate messages
Messages can sometimes be delivered more than once by the broker. When this happens,
the unique constraint on msg_id is violated, and we get errors. This change converts
the plain INSERT into a INSERT ... SELECT that will not insert anything if a message
with a matching msg_id already exists in the database. This will allow datanommer to
handle duplicate messages gracefully.