use INSERT ... SELECT when creating Users and Packages
A race condition exists in the get_or_create() method between checking for
the existence of a User or Package with a given name, and creating the entry
with that name. In the event that multiple instance of datanommer are
populating the same database with messages, this can lead to unique
constraint violations.

Convert get_or_create() to use INSERT ... SELECT to avoid this race condition.