Set inline=True when using INSERT ... FROM SELECT.
By default, sqlalchemy will try to retrieve the last inserted value
from an insert/update/delete. If doing an INSERT/SELECT, this can
results in either no values, or multiple values being inserted, and
this will break the retrieval. Setting inline=True disables the
retrieval.

See:

https://github.com/zzzeek/sqlalchemy/blob/rel_0_9/lib/sqlalchemy/sql/dml.py#L504

for a longer explanation. Note that this is only a problem in sqlalchemy
0.9. 1.0 and higher set inline=True automatically for all relevant
cases.