Bring back the users and packages tables
The Array type that we wanted to replace them with can become too big
for PostgreSQL's indexes (think of the list of packages in a message
triggered by a mass rebuild).

We got the following exception on insert:
`psycopg2.OperationalError: index row requires 1228104 bytes, maximum size is 8191`

Bring back the two tables, but don't use foreign keys pointing to the
messages table as this is not allowed on hypertables. That's why the
relationship needs to be declared with an explicit `primaryjoin`
argument.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>